US3025: Retrieve cartridge info with app
