+++ controller/app/controllers/emb_cart_events_controller.rb
@@ -1,7 +1,6 @@
 class EmbCartEventsController < BaseController
   respond_to :xml, :json
   before_filter :authenticate, :check_version
-  include LegacyBrokerHelper
 
   # POST /domain/[domain_id]/applications/[application_id]/cartridges/[cartridge_id]/events
   def create
