+++ controller/app/controllers/gears_controller.rb
@@ -1,7 +1,6 @@
 class GearsController < BaseController
   respond_to :xml, :json
   before_filter :authenticate, :check_version
-  include LegacyBrokerHelper
   
   def show
     domain_id = params[:domain_id]
