+++ controller/app/controllers/keys_controller.rb
@@ -1,7 +1,6 @@
 class KeysController < BaseController
   respond_to :xml, :json
   before_filter :authenticate, :check_version
-  include LegacyBrokerHelper
 
   #GET /user/keys
   def index
