+++ controller/app/controllers/legacy_broker_controller.rb
@@ -5,6 +5,7 @@ class LegacyBrokerController < ApplicationController
   rescue_from Exception, :with => :exception_handler
   include LegacyBrokerHelper
   include UserActionLogger
+  include CartridgeHelper
   
   # Initialize domain/app variables to be used for logging in user_action.log
   # The values will be set in the controllers handling the requests
