+++ controller/app/models/rest_application10.rb
@@ -1,5 +1,6 @@
 class RestApplication10 < OpenShift::Model
-  attr_accessor :framework, :creation_time, :uuid, :embedded, :aliases, :name, :gear_count, :links, :domain_id, :git_url, :app_url, :ssh_url, :gear_profile, :scalable, :health_check_path, :scale_min, :scale_max
+  attr_accessor :framework, :creation_time, :uuid, :embedded, :aliases, :name, :gear_count, :links, :domain_id, :git_url, 
+                :app_url, :ssh_url, :gear_profile, :scalable, :health_check_path, :scale_min, :scale_max, :cartridges
   include LegacyBrokerHelper
 
   def initialize(app, url, nolinks=false)
