+++ controller/app/models/rest_application12.rb
@@ -1,5 +1,6 @@
 class RestApplication12 < OpenShift::Model
-  attr_accessor :framework, :creation_time, :uuid, :embedded, :aliases, :name, :gear_count, :links, :domain_id, :git_url, :app_url, :ssh_url, :gear_profile, :scalable, :health_check_path, :building_with, :building_app, :build_job_url
+  attr_accessor :framework, :creation_time, :uuid, :embedded, :aliases, :name, :gear_count, :links, :domain_id, :git_url, :app_url, 
+                :ssh_url, :gear_profile, :scalable, :health_check_path, :building_with, :building_app, :build_job_url, :cartridges
   include LegacyBrokerHelper
 
   def initialize(app, url, nolinks=false)
